PSF Segment Stitching Demo
The scripts in this zip folder enable reconstructing the PSF of a diffuser using a sensor area smaller than that of the diffuser. 

3 Scripts:
ZemaxSimPSFSegments --> Zemax Interactive Code.  Generates overlapping PSF measurements acquired by a small sensor.
StitchPSFsegments - Matlab script that stitches the segmented PSF images together to recreate the total PSF of the diffuser.
PSFReconQuality - Plots the reconstruction quality as a function of the diffuser-to-sensor length ratio. The metric is simply a dot product between a standard PSF (no stitching) and a reconstructed (stitched) PSF.